+++ vignettes/precompile.R
@@ -3,3 +3,6 @@
 
 knitr::knit("vignettes/usage_prioritization.Rmd.orig", "vignettes/usage_prioritization.Rmd")
 knitr::knit("vignettes/getboundary.Rmd.orig", "vignettes/getboundary.Rmd")
+
+fs::dir_copy("figure", "vignettes/figure", overwrite = TRUE)
+fs::dir_delete("figure")